requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pandas==2.2.3
scikit-learn==1.6.1
tensorflow==2.16.2
keras==3.9.2
keras-tuner==1.4.7
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2023.3
requests==2.32.3
boto3~=1.38.16
pydantic~=2.11.4
python-dotenv~=1.1.0
mlflow==2.8.1
pyarrow==14.0.1